+++ packages/isaacscript-common/src/cachedClasses.ts
@@ -37,3 +37,27 @@ export const musicManager = MusicManager();
  * `Room`) is not safe and can lead to the game crashing in certain situations.
  */
 export const sfxManager = SFXManager();
+
+/**
+ * An object containing all 7 vanilla fonts that are pre-loaded and ready to use.
+ *
+ * For more information on the vanilla fonts and to see what they look like, see:
+ * https://wofsauge.github.io/IsaacDocs/rep/tutorials/Tutorial-Rendertext.html
+ */
+export const fonts = {
+  droid: Font(),
+  pfTempestaSevenCondensed: Font(),
+  teamMeatFont10: Font(),
+  teamMeatFont12: Font(),
+  teamMeatFont16Bold: Font(),
+  terminus: Font(),
+  upheaval: Font(),
+} as const;
+
+fonts.droid.Load("font/droid.fnt");
+fonts.pfTempestaSevenCondensed.Load("font/pftempestasevencondensed.fnt");
+fonts.teamMeatFont10.Load("font/teammeatfont10.fnt");
+fonts.teamMeatFont12.Load("font/teammeatfont12.fnt");
+fonts.teamMeatFont16Bold.Load("font/teammeatfont16bold.fnt");
+fonts.terminus.Load("font/terminus.fnt");
+fonts.upheaval.Load("font/upheaval.fnt");